# ir: add faster serializer