This Serializer which is implemented external to the IR node definition uses a StringBuilder to achieve about a 1.7x performance improvement when serializing.

Eventually, all implementations of the `serialize` methd should be replaced with a call to `Serializer.serialize`.

However, for this PR we keep the old code in place in order to allow for easy regression testing with the benchmark JAR like this:

```.sh
> java -cp utils/bin/firrtl-benchmark.jar  firrtl.benchmark.hot.SerializationBenchmark   ~/benchmarks/medium.pb 2 5 test
```

The two instances where `Serializer.serialize` intentionally differs from the old behavior are:

1. new lines without trailing white spaces (the old `indent` function indents empty lines)
2. attach statement serialization includes the serialization of the info node

### Hot Benchmarking

My hot benchmark results (with 10x warmup, 10x benchmark):

#### JVM 14

|                |  `[...].serialize`       |  `Serializer.serialize([...])` |
|----------|------------------|------------------------|
| small       |  640ms ± 5ms     | 425ms ± 5ms |
| medium  | 1307ms ± 101ms  | 818ms ± 28ms     |   


#### JVM 1.8

|                |  `[...].serialize`       |  `Serializer.serialize([...])` |
|----------|------------------|------------------------|
| small       |  1013ms ± 22ms     | 443ms ± 11ms |
| medium  | 2256ms ± 124ms  | 850ms ± 10ms     |   

_Caveat_: I have not independently evaluated the performance when serializing small chunks of work like individual expressions.

_Note_: https://github.com/freechipsproject/firrtl/pull/1690 brings about another 2x improvement on top of this for serialization performance.

### Contributor Checklist

- [x] Did you add Scaladoc to every public function/method?
- Did you update the FIRRTL spec to include every new feature/behavior?
  - not applicable
- Did you add at least one test demonstrating the PR?
  - besides the regression test that I performed locally by comparing the serialization of a large circuit, the old tests should cover this once we change the default serializer since the expected behavior stays the same
- [x] Did you address all `TODO` comments in the code?
- [x] Did you delete any extraneous printlns/debugging code?
- [x] Did you specify the type of improvement?
- [x] Did you state the API impact?
- [x] Did you specify the code generation impact?
- [x] Did you request a desired merge strategy?
- [x] Did you add text to be included in the Release Notes for this change?

#### Type of Improvement
- performance improvement

#### API Impact

- adds new `firrtl.ir.Serializer` object with one public `serialize` method and two public fields (`NewLine` and `Indent`)

#### Backend Code Generation Impact

- this should have no impact

#### Desired Merge Strategy

- squash

#### Release Notes
- new fast IR to String serialization with `firrtl.ir.Serializer.serialize(...)`

### Reviewer Checklist (only modified by reviewer)
- [ ] Did you add the appropriate labels?
- [ ] Did you mark the proper milestone (1.2.x, 1.3.0, 1.4.0) ?
- [ ] Did you review?
- [ ] Did you check whether all relevant Contributor checkboxes have been checked?
- [ ] Did you mark as `Please Merge`?