@ekiwi-sifive 2020-06-18
Some Node count statistics:

This only includes the nodes that would be serialized (- Info nodes), e.g. it ignores expression types. Also, these results are for HighForm/ Chirrtl

**small**
```
Reference: 422878
DefNode: 210591
DoPrim: 208438
IntWidth: 179616
SubField: 141089
UIntLiteral: 126690
SubIndex: 90428
eq: 69341
and: 64323
Field: 54251
UIntType: 53248
Default: 48793
Conditionally: 38257
EmptyStmt: 37616
Connect: 30059
Block: 21094
or: 20324
asUInt: 18304
Print: 18018
bits: 14505
BundleType: 13080
IsInvalid: 12403
DefWire: 10250
Flip: 5458
cat: 5407
PartialConnect: 4259
Mux: 3573
andr: 3103
orr: 3075
Stop: 2880
neq: 2435
asSInt: 2019
DefRegister: 1630
VectorType: 1106
Port: 1060
```

**medium**

```
Reference: 966780
DefNode: 485991
DoPrim: 477957
SubField: 458989
IntWidth: 362223
UIntLiteral: 258956
Connect: 111944
Field: 107218
eq: 104365
and: 103906
UIntType: 103773
Default: 99155
or: 84446
SubIndex: 81990
bits: 67106
Conditionally: 33729
EmptyStmt: 32871
cat: 27423
Block: 24908
BundleType: 24099
Mux: 20682
asSInt: 18832
IsInvalid: 17749
Print: 16557
asUInt: 15781
DefWire: 14288
Stop: 10252
Flip: 8063
xor: 7893
shr: 7598
orr: 6164
leq: 6135
cvt: 6011
DefRegister: 5314
andr: 4048
```