@ekiwi-sifive 2020-06-18
For LowForm (ignoring any custom passes and annotations that this benchmark should probably run though):

_Also, again, ignoring all `Info` and all expression types_

**small*

```
Reference: 374808
DoPrim: 256518
DefNode: 242874
IntWidth: 212039
UIntLiteral: 204733
and: 128019
eq: 61879
ValidIf: 30854
asUInt: 18304
Print: 17868
bits: 14347
or: 13874
Mux: 12998
Connect: 7923
UIntType: 7306
SubField: 6418
cat: 5407
Port: 4688
andr: 3103
orr: 3075
Stop: 2880
DefRegister: 2843
Input: 2807
asSInt: 2019
Output: 1881
xor: 977
leq: 850
neq: 835
not: 810
cvt: 556
tail: 421
dshl: 401
pad: 347
DefInstance: 321
sub: 269
```

**medium**

```
Reference: 1018754
DefNode: 582182
DoPrim: 547378
IntWidth: 341629
UIntLiteral: 304780
and: 186622
eq: 112923
or: 80323
SubField: 67859
bits: 60097
Mux: 58539
Connect: 52960
UIntType: 36715
Port: 25158
cat: 24680
asSInt: 17761
Print: 15425
asUInt: 15280
Input: 14587
DefRegister: 11357
Output: 10571
Stop: 9528
ValidIf: 8003
xor: 7301
shr: 6714
leq: 5935
cvt: 5674
orr: 4909
andr: 3942
not: 2774
tail: 1763
shl: 1573
add: 1508
dshl: 1493
neq: 1270
```