@jackkoenig 2020-06-18
> Some Node count statistics:

Thanks for collecting this information, it's very interesting. We should probably pick an order for each subtype informed by this. It probably doesn't matter that much but I think at least putting the types that are clearly the most prevalent at the top of each serialization function will have a small impact. I think if we wanted to pick the "best" order we'd need to include more inputs and do a more thorough investigation, but barring that we can definitely at least put `DefNode` at the top of `Statement`, and `Reference`, `DoPrim`, `SubField` at the top of `Expression`