@jackkoenig 2020-06-18
It would also be helpful to sort by group (eg. `Statement`, `Expression`), but even as is I think it's enough to at least pick the top few for each type