@ekiwi-sifive 2020-06-18
Yeah, I am doing a coarse sort. Some things depend a lot on Chirrtl vs LowForm, e.g. after compilation there are very few blocks and empty statements.