@jackkoenig 2020-06-18
> after compilation there are very few blocks and empty statements.

Yeah, most of those are probably part of `Conditionally`, nested `Block.mapStmt` now flattens out nested Blocks and EmptyStmts