@ekiwi-sifive 2020-06-18
So with the latest change:

medium@JVM1.8: 850ms ± 10ms -> 802ms ± 5ms -> 782ms ± 6ms
medium@JVM14: 818ms ± 28ms -> 720ms ± 13ms -> 707ms ± 32ms

But to be honest, we might be staring at noise here since even with the warmup (10) and repeated bench marking (10)  I get very different numbers between multiple runs, e.g. 827ms, 782ms, 730ms over three runs of the medium benchmark on JVM1.8.

I think it is time to stop micro-optimizing.