@jackkoenig 2020-06-24
Before we merge, is this something that should be backported? It could be used to speed some stuff up. @seldridge any thoughts?