@seldridge 2020-06-24
It's a pure API expansion, so I think it's reasonable to backport to 1.2.x. There may be conflicts with the WIR merge, but those are fixable.