@ucbjrl 2020-06-24
I don't think this should be backported. You want 1.2 to die off. You don't want to have to support multiple development branches. You don't have the resources.