@ekiwi-sifive 2020-06-24
This is also completely useless without a followup PR that replaces all implementations of `serialize` with a call to the new Serializer.