@jackkoenig 2020-06-24
We could backport to just 1.3.x to get the serialization speedups (with the follow on PR of course). But yeah I'm inclined to agree with @ucbjrl that we should maybe only be backporting bugfixes to `1.2.x`.