@seldridge 2020-06-24
Sounds good. I'm fine with a policy of backporting to 1.2.x only bug fixes. Greedily backporting is probably overzealous.