@ekiwi-sifive 2020-06-30
I just updated this PR in order to serialize the new `Verification` nodes.